rename: UserRepository 중복파일 삭제 및 JwtFilter 파일 위치 조정
