--- src/main/java/site/sonisori/sonisori/auth/jwt/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package site.sonisori.sonisori.auth.jwt.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import site.sonisori.sonisori.entity.User;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-	User findByUsername(String username);
-}
